requirements.txt
aiohttp==3.7.2
APScheduler==3.6.3
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
decorator==4.4.2
dnspython==1.16.0
greenlet==0.4.17
idna==2.10
multidict==5.0.0
pycparser==2.20
python-engineio==3.13.2
python-socketio==4.6.0
pytz==2020.4
requests==2.25.0
six==1.15.0
tornado==6.1
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.1
websocket-client==0.57.0
websockets==8.1
yarl==1.6.2
